requirements.txt
flask
numpy
opencv-python
Pillow
ultralytics
pyautogui
pytesseract
